requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
config==0.5.0.post0
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap3==14.2.0
django-filter==2.4.0
django-fontawesome-5==1.0.18
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
Pillow==8.0.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pyuploadcare==2.7.0
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==5.2.0
